Fix build warning in gc-alloc-profiler.cpp: cast to (size_t) (#44180)
`std::vector::size()` returns a `size_t`, so we need to cast the int
`jl_n_threads` to a `size_t` for the comparison.

(cherry picked from commit 15dcd5bb9c3458d52eb70dc59d685af399b54189)

+++ src/gc-alloc-profiler.cpp
@@ -71,7 +71,7 @@ extern "C" {  // Needed since these functions doesn't take any arguments.
 
 JL_DLLEXPORT void jl_start_alloc_profile(double sample_rate) {
     // We only need to do this once, the first time this is called.
-    while (g_alloc_profile.per_thread_profiles.size() < jl_n_threads) {
+    while (g_alloc_profile.per_thread_profiles.size() < (size_t)jl_n_threads) {
         g_alloc_profile.per_thread_profiles.push_back(jl_per_thread_alloc_profile_t{});
     }
